requirements.txt
Flask==3.0.1
python-dotenv==1.0.1
pika==1.3.2
pytz==2024.2
gunicorn==23.0.0
psycopg[binary,pool]
colorlog==6.9.0